+++ frontend/src/features/exams/components/TeacherQuestionWorkspace.jsx
@@ -5,6 +5,7 @@ import EmptyState from "../../../components/common/EmptyState";
 import QuestionCard from "./QuestionCard";
 import QuestionForm from "./QuestionForm";
 import QuestionImportPanel from "./QuestionImportPanel";
+import QuestionImportResources from "./QuestionImportResources";
 import {
   buildQuestionSummaryItems,
   filterQuestionItems,
@@ -144,6 +145,8 @@ export default function TeacherQuestionWorkspace({
           ))}
         </div>
       </Card>
+
+      <QuestionImportResources />
 
       <div
         className={